Added support for customizing an indicator view. Defaults to the standard typingIndicatorView
By calling -registerClassForIndicatorView: during initialization of the view controller, you can customize the appearance and behavior of the typing indicator view.

If not called, or a subclass of SLKTypingIndicatorView is passed, -typingIndicatorView will return the typing indicator view as before. If some other subclass of UIView is passed, typingIndicatorView will return nil, and you need to cast the return value of indicatorView to the appropriate type in you need to interact with it directly. Setting its hidden property to either true or false, will automatically animate.