leftbutton: fixed text-only button not displayed
When using only a label for the left button it previously wasn't displayed at all.
This was due to constraints recalculation was only done when the image is changed but not when the label is changed.
Now label is monitored for changes too and constraints are coreetly updated therefore displaying text-only buttons corretly.